+++ common/assets.go
@@ -9,6 +9,8 @@ const (
 	MicroUnit     = int64(1e6)
 
 	Update1Height = 100
+	// TODO: set to update height
+	Update2Height = 100
 )
 
 // GetCoinPercentage returns a certain percentage of coins.