+++ x/auth/module.go
@@ -23,6 +23,7 @@ import (
 
 	"github.com/certikfoundation/shentu/x/auth/client/cli"
 	"github.com/certikfoundation/shentu/x/auth/client/rest"
+
 	"github.com/certikfoundation/shentu/x/auth/keeper"
 	"github.com/certikfoundation/shentu/x/auth/simulation"
 	"github.com/certikfoundation/shentu/x/auth/types"