+++ x/cvm/keeper/querier_test.go
@@ -21,6 +21,7 @@ import (
 	"github.com/hyperledger/burrow/execution/evm/abi"
 	"github.com/hyperledger/burrow/txs/payload"
 
+	"github.com/certikfoundation/shentu/x/cvm/keeper"
 	. "github.com/certikfoundation/shentu/x/cvm/keeper"
 	"github.com/certikfoundation/shentu/x/cvm/types"
 )
@@ -92,7 +93,7 @@ func TestViewQuery(t *testing.T) {
 	getMyFavoriteNumberCall, _, err := abi.EncodeFunctionCall(
 		Hello55AbiJsonString,
 		"sayHi",
-		WrapLogger(ctx.Logger()),
+		keeper.WrapLogger(ctx.Logger()),
 	)
 	require.Nil(t, err)