+++ x/shield/keeper/invariants.go
@@ -14,7 +14,6 @@ func RegisterInvariants(ir sdk.InvariantRegistry, k Keeper) {
 	ir.RegisterRoute(types.ModuleName, "provider", ProviderInvariant(k))
 	ir.RegisterRoute(types.ModuleName, "shield", ShieldInvariant(k))
 	ir.RegisterRoute(types.ModuleName, "global-staking-pool", GlobalStakingPoolInvariant(k))
-	ir.RegisterRoute(types.ModuleName, "original-global-staking", StakingForShieldPurchaseInvariant(k))
 }
 
 // ModuleAccountInvariant checks that the module account coins reflects the sum of
@@ -122,24 +121,5 @@ func GlobalStakingPoolInvariant(keeper Keeper) sdk.Invariant {
 			fmt.Sprintf("\n\tsum of staked amount:  %s"+
 				"\n\tglobal staking pool amount: %s\n",
 				stakedInt, globalStakingPool.String())), broken
-	}
-}
-
-// StakingForShieldPurchaseInvariant checks that sum of original staked shield equals to the total.
-func StakingForShieldPurchaseInvariant(keeper Keeper) sdk.Invariant {
-	return func(ctx sdk.Context) (string, bool) {
-		globalStakingPool := keeper.GetGlobalShieldStakingPool(ctx)
-
-		sum := sdk.ZeroInt()
-		for _, os := range keeper.GetAllOriginalStakings(ctx) {
-			sum = sum.Add(os.Amount)
-		}
-
-		broken := !globalStakingPool.Equal(sum)
-
-		return sdk.FormatInvariant(types.ModuleName, "global-staking-pool",
-			fmt.Sprintf("\n\tsum of originally staked amount:  %s"+
-				"\n\tglobal staking pool amount: %s\n",
-				sum, globalStakingPool.String())), broken
 	}
 }