+++ CHANGELOG.md
@@ -46,8 +46,34 @@ Ref: https://keepachangelog.com/en/1.0.0/
 ### State Machine Breaking Changes
 ### Features
 ### Improvements
+* [\#230](https://github.com/certikfoundation/shentu/pull/230) Optimized shield invariant & removed crisis module from endblocker
+### Tests
 ### Bug Fixes
 
+## [v1.3.1] - 02-05-2021
+
+### Client Breaking Changes
+### API Breaking Changes
+### State Machine Breaking Changes
+### Features
+### Improvements
+* [\#230](https://github.com/certikfoundation/shentu/pull/230) Optimized shield invariant & removed crisis module from endblocker
+### Tests
+### Bug Fixes
+
+## [v1.3.0] - 01-15-2021
+
+### Client Breaking Changes
+### API Breaking Changes
+### State Machine Breaking Changes
+### Features
+### Improvements
+* [\#219](https://github.com/certikfoundation/shentu/pull/219) Remove internal sub-packages.
+* [\#177](https://github.com/certikfoundation/shentu/pull/177) Updated Swagger docs.
+### Tests
+* [\#219](https://github.com/certikfoundation/shentu/pull/180) Implement SimApp for testing.
+### Bug Fixes
+* [\#216](https://github.com/certikfoundation/shentu/pull/216) Fixed Shield fee distribution.
 
 ## [v1.2.0] - 11-20-2020